refactor: move prefs class to common
