+++ build-tools/build.gradle.kts
@@ -29,6 +29,7 @@ android {
 
 
 dependencies {
+    implementation(projects.common)
     implementation(projects.project)
     implementation(projects.util)
     implementation(projects.kotlinc)
